contacts_api/requirements.txt
fastapi==0.101.0
fastapi_limiter
cloudinary
python-dotenv
uvicorn
sqlalchemy
psycopg2
passlib
python-jose
pydantic[email]
python-multipart
pytest
httpx
aioredis
redis